+++ app/models/users/project_authorization_cache.rb
@@ -28,34 +28,42 @@
 # See docs/COPYRIGHT.rdoc for more details.
 #++
 
-class User::ProjectRoleCache
+class Users::ProjectAuthorizationCache
   attr_accessor :user
 
   def initialize(user)
     self.user = user
   end
 
-  def fetch(project)
-    cache[project] ||= roles(project)
+  def cache(actions)
+    cached_actions = if actions.is_a?(Array)
+                       actions
+                     else
+                       [actions]
+                     end
+
+    cached_actions.each do |action|
+      allowed_project_ids = Project.allowed_to(user, action).pluck(:id)
+
+      projects_by_actions[normalized_permission_name(action)] = allowed_project_ids
+    end
+  end
+
+  def cached?(action)
+    projects_by_actions[normalized_permission_name(action)]
+  end
+
+  def allowed?(action, project)
+    projects_by_actions[normalized_permission_name(action)].include? project.id
   end
 
   private
 
-  def roles(project)
-    # No role on archived projects
-    return [] unless !project || project&.active?
-
-    # Return all roles if user is admin
-    return all_givable_roles if user.admin?
-
-    ::Authorization.roles(user, project).eager_load(:role_permissions)
+  def normalized_permission_name(action)
+    OpenProject::AccessControl.permission(action)
   end
 
-  def cache
-    @cache ||= {}
-  end
-
-  def all_givable_roles
-    @all_givable_roles ||= Role.givable.to_a
+  def projects_by_actions
+    @projects_by_actions ||= {}
   end
 end
