+++ app/models/users/project_role_cache.rb
@@ -28,42 +28,34 @@
 # See docs/COPYRIGHT.rdoc for more details.
 #++
 
-class User::ProjectAuthorizationCache
+class Users::ProjectRoleCache
   attr_accessor :user
 
   def initialize(user)
     self.user = user
   end
 
-  def cache(actions)
-    cached_actions = if actions.is_a?(Array)
-                       actions
-                     else
-                       [actions]
-                     end
-
-    cached_actions.each do |action|
-      allowed_project_ids = Project.allowed_to(user, action).pluck(:id)
-
-      projects_by_actions[normalized_permission_name(action)] = allowed_project_ids
-    end
-  end
-
-  def cached?(action)
-    projects_by_actions[normalized_permission_name(action)]
-  end
-
-  def allowed?(action, project)
-    projects_by_actions[normalized_permission_name(action)].include? project.id
+  def fetch(project)
+    cache[project] ||= roles(project)
   end
 
   private
 
-  def normalized_permission_name(action)
-    OpenProject::AccessControl.permission(action)
+  def roles(project)
+    # No role on archived projects
+    return [] unless !project || project&.active?
+
+    # Return all roles if user is admin
+    return all_givable_roles if user.admin?
+
+    ::Authorization.roles(user, project).eager_load(:role_permissions)
   end
 
-  def projects_by_actions
-    @projects_by_actions ||= {}
+  def cache
+    @cache ||= {}
+  end
+
+  def all_givable_roles
+    @all_givable_roles ||= Role.givable.to_a
   end
 end
